# openapi/api_recipe.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Potion maker - OpenAPI 3.0
 *
 * Сервис изготовления зелий
 *
 * API version: 1.0.0
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// RecipeAPIController binds http requests to an api service and writes the service results to the http response
type RecipeAPIController struct {
	service      RecipeAPIServicer
	errorHandler ErrorHandler
}

// RecipeAPIOption for how the controller is set up.
type RecipeAPIOption func(*RecipeAPIController)

// WithRecipeAPIErrorHandler inject ErrorHandler into controller
func WithRecipeAPIErrorHandler(h ErrorHandler) RecipeAPIOption {
	return func(c *RecipeAPIController) {
		c.errorHandler = h
	}
}

// NewRecipeAPIController creates a default api controller
func NewRecipeAPIController(s RecipeAPIServicer, opts ...RecipeAPIOption) *RecipeAPIController {
	controller := &RecipeAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the RecipeAPIController
func (c *RecipeAPIController) Routes() Routes {
	return Routes{
		"RecipesList": Route{
			strings.ToUpper("Get"),
			"/api/v1/recipes",
			c.RecipesList,
		},
		"GetRecipeByUUID": Route{
			strings.ToUpper("Get"),
			"/api/v1/recipe/{uuid}",
			c.GetRecipeByUUID,
		},
		"UpdateRecipeByUUID": Route{
			strings.ToUpper("Put"),
			"/api/v1/recipe/{uuid}",
			c.UpdateRecipeByUUID,
		},
		"DeleteRecipeByUUID": Route{
			strings.ToUpper("Delete"),
			"/api/v1/recipe/{uuid}",
			c.DeleteRecipeByUUID,
		},
	}
}

// RecipesList - recipes list
func (c *RecipeAPIController) RecipesList(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.RecipesList(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetRecipeByUUID - Find recipe by UUID
func (c *RecipeAPIController) GetRecipeByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.GetRecipeByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateRecipeByUUID - Update a recipe by uuid
func (c *RecipeAPIController) UpdateRecipeByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	recipeParam := Recipe{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&recipeParam); err != nil && !errors.Is(err, io.EOF) {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertRecipeRequired(recipeParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertRecipeConstraints(recipeParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateRecipeByUUID(r.Context(), uuidParam, recipeParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteRecipeByUUID - Delete recipe
func (c *RecipeAPIController) DeleteRecipeByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.DeleteRecipeByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}