# openapi/api_witch.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Potion maker - OpenAPI 3.0
 *
 * Сервис изготовления зелий
 *
 * API version: 1.0.0
 */

package openapi

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// WitchAPIController binds http requests to an api service and writes the service results to the http response
type WitchAPIController struct {
	service      WitchAPIServicer
	errorHandler ErrorHandler
}

// WitchAPIOption for how the controller is set up.
type WitchAPIOption func(*WitchAPIController)

// WithWitchAPIErrorHandler inject ErrorHandler into controller
func WithWitchAPIErrorHandler(h ErrorHandler) WitchAPIOption {
	return func(c *WitchAPIController) {
		c.errorHandler = h
	}
}

// NewWitchAPIController creates a default api controller
func NewWitchAPIController(s WitchAPIServicer, opts ...WitchAPIOption) *WitchAPIController {
	controller := &WitchAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the WitchAPIController
func (c *WitchAPIController) Routes() Routes {
	return Routes{
		"WitchesList": Route{
			strings.ToUpper("Get"),
			"/api/v1/witches",
			c.WitchesList,
		},
		"GetWitchByUUID": Route{
			strings.ToUpper("Get"),
			"/api/v1/witch/{uuid}",
			c.GetWitchByUUID,
		},
		"UpdateWitchByUUID": Route{
			strings.ToUpper("Put"),
			"/api/v1/witch/{uuid}",
			c.UpdateWitchByUUID,
		},
		"DeleteWitchByUUID": Route{
			strings.ToUpper("Delete"),
			"/api/v1/witch/{uuid}",
			c.DeleteWitchByUUID,
		},
	}
}

// WitchesList - witches list
func (c *WitchAPIController) WitchesList(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.WitchesList(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetWitchByUUID - Find witch by UUID
func (c *WitchAPIController) GetWitchByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.GetWitchByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateWitchByUUID - Update a witch in the store with form data
func (c *WitchAPIController) UpdateWitchByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.UpdateWitchByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteWitchByUUID - Deletes a witch
func (c *WitchAPIController) DeleteWitchByUUID(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	uuidParam := params["uuid"]
	if uuidParam == "" {
		c.errorHandler(w, r, &RequiredError{"uuid"}, nil)
		return
	}
	result, err := c.service.DeleteWitchByUUID(r.Context(), uuidParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}